fix: alteração nas migrations
- alterando para que os valores sejam salvos sempre em centavos para melhor controle

+++ api/database/migrations/2026_05_22_000001_create_wallets_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
-            $table->decimal('balance', 15, 2)->default(0);
+            $table->unsignedBigInteger('balance')->default(0);
             $table->timestamps();
         });
     }
